+++ examples/humanoid_jvrc_com.py
@@ -5,10 +5,10 @@
 #
 # /// script
 # dependencies = ["daqp", "loop-rate-limiters", "meshcat", "pin-pink",
-# "qpsolvers", "robot_descriptions>=1.21"]
+# "qpsolvers", "robot_descriptions", "typing_extensions"]
 # ///
 
-"""G1 humanoid squat by regulating CoM."""
+"""JVRC-1 humanoid backbending by regulating its CoM position."""
 
 import meshcat.geometry as g
 import meshcat.transformations as tf
@@ -20,12 +20,12 @@ from robot_descriptions.loaders.pinocchio import load_robot_description
 
 import pink
 from pink import solve_ik
-from pink.tasks import ComTask, FrameTask, PostureTask
+from pink.tasks import ComTask, FrameTask
 from pink.visualization import start_meshcat_visualizer
 
 if __name__ == "__main__":
     robot = load_robot_description(
-        "g1_description", root_joint=pin.JointModelFreeFlyer()
+        "jvrc_description", root_joint=pin.JointModelFreeFlyer()
     )
 
     # Initialize visualization
@@ -37,49 +37,36 @@ if __name__ == "__main__":
         g.Sphere(0.03), g.MeshLambertMaterial(color=0xFF0000)
     )
 
-    q_ref = np.zeros(robot.nq)
-    q_ref[2] = 0.72
-    q_ref[6] = 1.0
+    configuration = pink.Configuration(robot.model, robot.data, robot.q0)
+    viz.display(configuration.q)
 
-    configuration = pink.Configuration(robot.model, robot.data, q_ref)
-    pelvis_orientation_task = FrameTask(
-        "pelvis",
-        position_cost=0.0,  # [cost] / [m]
-        orientation_cost=10.0,  # [cost] / [rad]
+    left_ankle_task = FrameTask(
+        "l_ankle", position_cost=1.0, orientation_cost=3.0
+    )
+    right_ankle_task = FrameTask(
+        "r_ankle", position_cost=1.0, orientation_cost=3.0
+    )
+    com_task = ComTask(cost=2.0)
+
+    transform_l_ankle_target_to_init = pin.SE3(
+        np.eye(3), np.array([0.1, 0.0, 0.0])
+    )
+    transform_r_ankle_target_to_init = pin.SE3(
+        np.eye(3), np.array([-0.1, 0.0, 0.0])
+    )
+    left_ankle_task.set_target(
+        configuration.get_transform_frame_to_world("l_ankle")
+        * transform_l_ankle_target_to_init
+    )
+    right_ankle_task.set_target(
+        configuration.get_transform_frame_to_world("r_ankle")
+        * transform_r_ankle_target_to_init
     )
 
-    com_task = ComTask(cost=200.0)
-    com_task.set_target_from_configuration(configuration)
+    initial_com = pin.centerOfMass(robot.model, robot.data, configuration.q)
+    com_task.set_target(initial_com)
 
-    posture_task = PostureTask(
-        cost=1e-1,  # [cost] / [rad]
-    )
-
-    tasks = [pelvis_orientation_task, posture_task, com_task]
-
-    for foot in ["right_ankle_roll_link", "left_ankle_roll_link"]:
-        task = FrameTask(
-            foot,
-            position_cost=[2.0, 2.0, 200],  # [cost] / [m]
-            orientation_cost=10.0,  # [cost] / [rad]
-        )
-        tasks.append(task)
-
-    for arm_points in ["right_wrist_yaw_link", "left_wrist_yaw_link"]:
-        task = FrameTask(
-            arm_points,
-            position_cost=4.0,  # [cost] / [m]
-            orientation_cost=0.0,  # [cost] / [rad]
-        )
-        tasks.append(task)
-
-    for task in tasks:
-        task.set_target_from_configuration(configuration)
-        if isinstance(task, FrameTask):
-            target = task.transform_target_to_world
-            if task.frame in ["right_wrist_yaw_link", "left_wrist_yaw_link"]:
-                target.translation += np.array([-0.1, 0.0, -0.2])
-                task.set_target(target)
+    tasks = [com_task, left_ankle_task, right_ankle_task]
 
     # Select QP solver
     solver = qpsolvers.available_solvers[0]
@@ -97,10 +84,11 @@ if __name__ == "__main__":
     while True:
         pin.centerOfMass(robot.model, robot.data, configuration.q)
         com = robot.data.com[0]
+
         # Update CoM target
         Az = 0.05
-        desired_com = np.zeros(3)
-        desired_com[2] = 0.55 + Az * np.sin(omega * t)
+        desired_com = initial_com.copy()
+        desired_com[2] += Az * np.sin(omega * t)
         com_task.set_target(desired_com)
 
         velocity = solve_ik(
@@ -108,7 +96,6 @@ if __name__ == "__main__":
             tasks,
             dt,
             solver=solver,
-            damping=0.01,
             safety_break=False,
         )
         configuration.integrate_inplace(velocity, dt)